+++ src/main/java/org/project/openbaton/nubomedia/api/core/Main.java
@@ -1,4 +1,4 @@
-package org.project.openbaton.nubomedia.api;
+package org.project.openbaton.nubomedia.api.core;
 
 import org.project.openbaton.nubomedia.api.openshift.beans.OpenshiftConfiguration;
 import org.springframework.boot.SpringApplication;
