+++ src/main/java/org/project/openbaton/nubomedia/api/core/OpenbatonManager.java
@@ -1,4 +1,4 @@
-package org.project.openbaton.nubomedia.api;
+package org.project.openbaton.nubomedia.api.core;
 
 import org.openbaton.catalogue.mano.common.*;
 import org.openbaton.catalogue.mano.descriptor.InternalVirtualLink;
