+++ src/main/java/org/project/openbaton/nubomedia/api/core/SimpleCorsFilter.java
@@ -1,4 +1,4 @@
-package org.project.openbaton.nubomedia.api;
+package org.project.openbaton.nubomedia.api.core;
 
 import org.springframework.core.Ordered;
 import org.springframework.core.annotation.Order;
