+++ src/main/java/org/project/openbaton/nubomedia/api/openshift/beans/SecretManager.java
@@ -1,8 +1,7 @@
 package org.project.openbaton.nubomedia.api.openshift.beans;
 
 import com.google.gson.Gson;
-import org.mockito.internal.util.collections.ArrayUtils;
-import org.project.openbaton.nubomedia.api.openshift.MessageBuilderFactory;
+import org.project.openbaton.nubomedia.api.openshift.builders.MessageBuilderFactory;
 import org.project.openbaton.nubomedia.api.openshift.exceptions.UnauthorizedException;
 import org.project.openbaton.nubomedia.api.openshift.json.*;
 import org.slf4j.Logger;
